requirements.txt
black==20.8b1
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-heroku==0.0.0
flake8==3.8.4
gunicorn==20.1.0
idna==2.6
isort==5.6.4
Jinja2==2.11.2
mecab-python3==1.0.3
numpy==1.20.2
pandas==1.2.4
scipy==1.6.3
tweepy==3.10.0
unidic-lite==1.0.8
whitenoise==5.3.0
psycopg2>=2.8,<2.9
ipadic==1.0.0
pymlask==0.3.2
